split up SessionTest and nullable-ize it
